完成 OnlineAvgFunction, 有待测试

+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/util/OnlineAggregationUtils.scala
@@ -38,7 +38,7 @@ object commonMath {
   // 根据 errorBound, samplesCount 和 T 计算置信度 P,
   // T 需要在对应的聚合函数里增量求得.
   def calcConfidence(errorBound: Double,
-                     samplesCount: Int,
+                     samplesCount: Long,
                      T_Denominator: Double): Double = {
     2 * CNDF((errorBound * math.sqrt(samplesCount)) / T_Denominator) - 1
   }
